backend/requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
python-decouple==3.6
pytest
pytest-django
requests
pytest-bdd
stripe==13.1.1